+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -330,7 +330,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTRJBCC6BQP4SER6PMN6CGF3"
+        "01KTRMH8EBNW0ZBX7KTP01D201"
     )
     r"""The unique identifier of the data source"""