+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTRJBCC59NY09V2E6YJP9WE2"
+        "01KTRMH8EA1PKYVR01AXHK5A7F"
     )
     r"""The unique identifier of the data source"""