+++ packages/orq-rc/RELEASES.md
@@ -8599,3 +8599,13 @@ Based on:
 - [python v4.11.0-rc.9] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.9] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.9 - packages/orq-rc
+
+## 2026-06-10 11:28:40
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.774.1 (2.899.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.10] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.10] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.10 - packages/orq-rc